Move exceptions to binary-core-api module
