+++ library/binary-core-api/src/nonJvmMain/kotlin/io/matthewnelson/kmp/tor/binary/core/api/-NonJvmPlatform.kt
@@ -15,7 +15,7 @@
  **/
 @file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 
-package io.matthewnelson.kmp.tor.binary.core
+package io.matthewnelson.kmp.tor.binary.core.api
 
 public actual open class IOException: Exception {
     public actual constructor(): super()
