+++ library/binary-core/src/mingwMain/kotlin/io/matthewnelson/kmp/tor/binary/core/internal/MingwFs.kt
@@ -17,8 +17,8 @@
 
 package io.matthewnelson.kmp.tor.binary.core.internal
 
-import io.matthewnelson.kmp.tor.binary.core.FileNotFoundException
-import io.matthewnelson.kmp.tor.binary.core.IOException
+import io.matthewnelson.kmp.tor.binary.core.api.FileNotFoundException
+import io.matthewnelson.kmp.tor.binary.core.api.IOException
 import io.matthewnelson.kmp.tor.binary.core.InternalKmpTorBinaryApi
 import kotlinx.cinterop.*
 import platform.posix.*
